:sparkles: add usuario service e usuario controller

+++ src/service/ProdutoService.ts
@@ -33,15 +33,21 @@ export class ProdutoService {
     return produto;
   }
 
-  async deletarProduto(produtoData: any): Promise<Produto> {
-    const { id } = produtoData;
-
-    const produto = new Produto(id, "", 0, "", "", 0);
-
+  async deletarProduto(produtoData: any): Promise<Produto[]> {
+    const { id, nome, preco, descricao, imagem, quantidade, usuario_id } = produtoData;
+  
+    const produto = new Produto(id, nome, preco, descricao, imagem, quantidade, usuario_id);
+    const produtoEncontrado = await this.productRepository.buscaProdutoPorIDeNome(id, nome);
+  
+    if (produtoEncontrado.length === 0) {
+      throw new Error("404!!! Not Found - Produto não encontrado!!");
+    }
+  
     await this.productRepository.deleteProduto(produto);
     console.log("Service - Produto deletado: ", produto);
-    return produto;
+    return produtoEncontrado;
   }
+  
 
   async filtrarProduto(id: number, usuario_id: number): Promise<Produto> {
     const produto = await this.productRepository.filterProduto(id, usuario_id);
